# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hamboom API",
    "version": "0.1.0",
    "description": "REST APIِ پلتفرمِ وایت‌بوردِ همکاریِ هم‌بوم (ماژول M3)."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "Error": {
        "description": "خطا (قالبِ یکسانِ apiError)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "VALIDATION_ERROR",
                  "UNAUTHORIZED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "CONFLICT",
                  "RATE_LIMITED",
                  "INTERNAL",
                  "USER_NOT_FOUND",
                  "TEAM_NOT_FOUND",
                  "BOARD_NOT_FOUND",
                  "BOARD_ID_MALFORMED",
                  "OTP_INVALID",
                  "OTP_EXPIRED",
                  "OTP_TOO_MANY",
                  "TOKEN_REUSED"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "propertyNames": {
                  "type": "string"
                },
                "additionalProperties": {}
              },
              "requestId": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "requestId"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "OtpRequestBody": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "pattern": "^09\\d{9}$"
          }
        },
        "required": [
          "phone"
        ],
        "additionalProperties": false
      },
      "OtpVerifyBody": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "pattern": "^09\\d{9}$"
          },
          "code": {
            "type": "string",
            "pattern": "^\\d{6}$"
          }
        },
        "required": [
          "phone",
          "code"
        ],
        "additionalProperties": false
      },
...
      "PutAccessBody": {
        "type": "object",
        "properties": {
          "accessMode": {
            "type": "string",
            "enum": [
              "private",
              "team",
              "link_view",
              "link_edit"
            ]
          },
          "regenerate": {
            "type": "boolean"
          }
        },
        "required": [
          "accessMode"
        ],
        "additionalProperties": false
      },
      "ResolveLinkBody": {
        "type": "object",
        "properties": {
          "linkToken": {
            "type": "string",
            "minLength": 10
          }
        },
        "required": [
          "linkToken"
        ],
        "additionalProperties": false
      },
      "AddBoardMemberBody": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "editor",
              "viewer"
            ]
          }
        },
        "required": [
          "userId",
          "role"
        ],
        "additionalProperties": false
      },
      "PatchBoardMemberRoleBody": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "editor",
              "viewer"
            ]
          }
        },
        "required": [
          "role"
        ],
        "additionalProperties": false
      }
    }
  },
  "paths": {
...
    "/auth/otp/request": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "ارسالِ OTP (ضدِ enumeration، rate-limit)",
        "responses": {
          "200": {
            "description": "موفق"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequestBody"
              }
            }
          }
        }
      }
    },
    "/auth/otp/verify": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "تاییدِ OTP → accessToken + کوکیِ refresh",
        "responses": {
          "200": {
            "description": "موفق"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyBody"
              }
            }
          }
        }
      }
    },
...
    "/boards/{id}/access": {
      "get": {
        "tags": [
          "board-access"
        ],
        "summary": "حالتِ اشتراک + اعضا (viewer+)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "موفق"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "put": {
        "tags": [
          "board-access"
        ],
        "summary": "تنظیمِ حالت + تولید/ابطالِ لینک (owner)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "موفق"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutAccessBody"
              }
            }
          }
        }
      }
    },
    "/public/boards/resolve": {
      "post": {
        "tags": [
          "board-access"
        ],
        "summary": "resolveِ مهمانِ لینک → گرنتِ ماندگار (DP-4)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "موفق"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveLinkBody"
              }
            }
          }
        }
      }
    },
    "/boards/{id}/members": {
      "post": {
        "tags": [
          "board-access"
        ],
        "summary": "افزودنِ عضوِ مستقیم (owner)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "موفق"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddBoardMemberBody"
              }
            }
          }
        }
      }
    },
    "/boards/{id}/members/{userId}": {
      "patch": {
        "tags": [
          "board-access"
        ],
        "summary": "تغییرِ نقشِ عضوِ مستقیم (owner)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "موفق"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchBoardMemberRoleBody"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "board-access"
        ],
        "summary": "حذفِ عضوِ مستقیم (owner)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "بدونِ محتوا"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  }
}